+++ src/functions.js
@@ -30,7 +30,39 @@ const verifyToken = (token) => {
     }
 }
 
+const isAuthorized = (router, roles) =>{
+    router.use((req, res, next) => {
+        const token = req.get("Authorization")
+        var access_token = undefined
+    
+        if (typeof token !== "undefined"){
+            access_token = token
+        }else{
+            access_token = req.cookies.access_token
+        }
+    
+        if (typeof access_token !== "undefined"){
+            const token_decoded = verifyToken(access_token)
+            if (typeof token_decoded !== "undefined" && access_token !== "token expired"){
+                const { user, rol } = token_decoded
+                if (typeof user !== "undefined" && typeof rol !== "undefined"){
+                    const getRole = roles.find((item)=> item == rol)
+                    if (typeof getRole !== "undefined"){
+                        next()
+                        return
+                    }
+                }
+            }else{
+                res.status(401).send("Your token has expired. Please just re-log in and try again.")
+                return 
+            }
+        } 
+        res.status(403).send("You are not allowed to do this.")
+    })
+}
+
 export const tokensMethods = {
     signToken,
-    verifyToken
+    verifyToken,
+    isAuthorized
 }
