+++ src/routes/comentario.route.js
@@ -1,14 +1,18 @@
 import { Router } from "express";
 import { methods as comentarioMethods} from "./../controllers/comentarioController.js"
+import { tokensMethods } from "../functions.js";
 
 const router = Router();
 
-router.get("/:id", comentarioMethods.getComentarioById)
+tokensMethods.isAuthorized(router, ["docente", "administrador"])
 router.get("/docente/:cedula", comentarioMethods.getComentarioByDocente)
-router.get("/salon/:salon", comentarioMethods.getComentarioBySalon)
-router.get("/", comentarioMethods.getAllComentarios)
 router.post("/register", comentarioMethods.registerComentario)
-router.delete("/delete/:id", comentarioMethods.deleteComentarioById)
 router.delete("/delete/docente/:cedula", comentarioMethods.deleteComentarioByDocente)
+
+tokensMethods.isAuthorized(router, ["administrador"])
+router.get("/:id", comentarioMethods.getComentarioById)
+router.get("/", comentarioMethods.getAllComentarios)
+router.get("/salon/:salon", comentarioMethods.getComentarioBySalon)
+router.delete("/delete/:id", comentarioMethods.deleteComentarioById)
 
 export default router
