+++ src/routes/salon.route.js
@@ -1,7 +1,10 @@
 import { Router } from "express";
 import { methods as salonMethods} from "./../controllers/salonController.js"
+import { tokensMethods } from "../functions.js";
 
 const router = Router();
+
+tokensMethods.isAuthorized(router, ["administrador"])
 
 router.get("/:id", salonMethods.getSalonById)
 router.get("/", salonMethods.getSalones)
